requirements.txt
transmissionrpc
feedparser
requests
watchdog
jsonrpclib
psutil
six
pathtools
psycopg2
jinja2
flask
flask-login
flask-script
flask-sqlalchemy
flask-mail
flask-cache
flask-babel
-e git+git://github.com/seatgeek/fuzzywuzzy.git#egg=fuzzywuzzy
-e git+git://github.com/alberanid/imdbpy.git#egg=imdbpy
